feat: add global repo-root targeting and worktree handoff flow
