+++ src/commands/doctor/doctor.command.ts
@@ -12,6 +12,7 @@ type DoctorCommandOptions = {
   runtime?: boolean;
   portal?: boolean;
   osgi?: boolean;
+  repoRoot?: string;
 };
 
 export function createDoctorCommand(): Command {
