+++ src/index.ts
@@ -1,5 +1,6 @@
 #!/usr/bin/env node
 import {normalizeCliError, resolveOutputFormatFromArgv, toCliErrorPayload} from './cli/errors.js';
+import {resolveCommandRoot} from './cli/command-context.js';
 import {createCli} from './cli/create-cli.js';
 import {buildContextualRootSummary} from './cli/contextual-help.js';
 import {sanitizeErrorMessage} from './core/errors-sanitize.js';
@@ -7,7 +8,7 @@ import {sanitizeErrorMessage} from './core/errors-sanitize.js';
 async function main(): Promise<void> {
   const cli = createCli();
   if (process.argv.length <= 2) {
-    process.stdout.write(`${buildContextualRootSummary(process.cwd())}\n`);
+    process.stdout.write(`${buildContextualRootSummary(resolveCommandRoot(undefined, process.argv, process.env))}\n`);
     return;
   }
   await cli.parseAsync(process.argv);
